+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.13.2 (2018-10-26)
+
+- Feature [Core] Upgrade digitalstate/core dependency to 0.13.2
+- Bug [Migration] Fix serialization bug for parameter and config migration
+- Bug [Tenant] Add missing business units and identity tenant loaders
+
 ## 0.13.1 (2018-10-19)
 
 - Bug [App] Fix racing issue related to `secret` app parameter being null early in the execution